fix duplicate binding because of unique overflow http://tapd.oa.com/TBase_Oracle_Migration/bugtrace/bugs/view?bug_id=1020421696094794629&jump_count=1 (merge request !1004)
Squash merge branch 'sigmalin_v5' into 'Tbase_v5.06'
fix duplicate binding because of unique overflow http://tapd.oa.com/TBase_Oracle_Migration/bugtrace/bugs/view?bug_id=1020421696094794629&jump_count=1

TAPD: --bug=094794629

(cherry picked from commit c3df25c4)

4064c9fe fix duplicate binding because of unique overflow http://tapd.oa.com/TBase_Oracle_Migration/bugtrace/bugs/view?bug_id=1020421696094794629&jump_count=1

+++ src/include/pgxc/planner.h
@@ -247,7 +247,7 @@ typedef struct
     bool         execOnAll;
     SimpleSort *sort;
     char       *cursor;
-    int            unique;
+    int64       unique;
 #ifdef __TBASE__
     /*
       * if gather is under remotesubplan, parallel worker can send tuples
